refactor: use shared generic dropdown util and spinner
Drop the local dropdown-utils in favour of generic's initSearchableDropdowns,
inject the shared control tokens from starter.js, and replace the progress-wheel
<img> usages (side panel, popup, bulk export) with the generic .sbb-spinner.
