+++ src/main/resources/webapp/pdf-exporter/html/popupForm.html
@@ -1,5 +1,5 @@
 <div class="form-wrapper">
-    <div class="in-progress-overlay"><img src='/polarion/ria/images/progressWheel48.svg' alt=""/><span id="in-progress-message"></span></div>
+    <div class="in-progress-overlay"><span class="sbb-spinner" role="img" aria-label="Loading" style="width:48px;height:48px;"></span><span id="in-progress-message"></span></div>
 
     <div class="notifications">
         <div class="alert alert-warning" style="display: none"></div>
