+++ src/main/resources/webapp/pdf-exporter/html/sidePanelContent.html
@@ -254,7 +254,7 @@
         <button type='button' id='export-pdf'>
             <img class='append-build-number' src='/polarion/ria/images/dle/operations/actionPdfExport16.svg' alt=''/>Export to PDF
         </button>
-        <img id='export-pdf-progress' src='/polarion/ria/images/progressWheel48.svg' alt=''/>
+        <span id='export-pdf-progress' class='sbb-spinner' role='img' aria-label='Loading'></span>
         <div id='export-error'></div>
         <div id='export-warning'></div>
     </div>
@@ -263,7 +263,7 @@
             <img class='append-build-number' src='/polarion/ria/images/dle/align_center.png' alt=''/>Validate pages width
         </button>
         <div id='validate-ok'></div>
-        <img id='validate-pdf-progress' src='/polarion/ria/images/progressWheel48.svg' alt=''/>
+        <span id='validate-pdf-progress' class='sbb-spinner' role='img' aria-label='Loading'></span>
         <br>
         <div id='validate-error'></div>
     </div>
