+++ src/test/js/path-redirector.mjs
@@ -1,5 +1,7 @@
-// This path redirector is used to replace paths in imports of JS modules from Generic project which is not available for tests
-// by paths to these modules in build target, in its generic-jar-content folder. See scripts/test item in package.json located in project's root
+// Generic modules aren't present in this project's source tree; they are built into
+// target/generic-jar-content. This loader redirects any import of a generic asset — whether written
+// as an absolute `/polarion/<ext>/ui/generic/js/...` path or a relative `../ui/generic/js/...` one —
+// to that folder. See the scripts/test entry in package.json.
 
 import { fileURLToPath, pathToFileURL } from 'url';
 import { resolve as pathResolve, dirname } from 'path';
@@ -7,18 +9,14 @@ import { resolve as pathResolve, dirname } from 'path';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-const pathMappings = {
-    '/polarion/pdf-exporter/ui/generic/js/modules/ConfigurationsPane.js': pathResolve(__dirname, '../../../target/generic-jar-content/js/modules/ConfigurationsPane.js'),
-    '/polarion/pdf-exporter/ui/generic/js/modules/CustomSelect.js': pathResolve(__dirname, '../../../target/generic-jar-content/js/modules/CustomSelect.js'),
-    '/polarion/pdf-exporter/ui/generic/js/modules/ExtensionContext.js': pathResolve(__dirname, '../../../target/generic-jar-content/js/modules/ExtensionContext.js'),
-    '/polarion/pdf-exporter/ui/generic/js/modules/SearchableDropdown.js': pathResolve(__dirname, '../../../target/generic-jar-content/js/modules/SearchableDropdown.js'),
-    '/polarion/pdf-exporter/ui/generic/js/micromodal.min.js': pathResolve(__dirname, '../../../target/generic-jar-content/js/micromodal.min.js')
-};
+const GENERIC_SUFFIX = /(?:^|\/)(?:ui\/)?generic\/(js\/.+\.js)$/;
 
 export async function resolve(specifier, context, nextResolve) {
-    if (pathMappings[specifier]) {
+    const match = GENERIC_SUFFIX.exec(specifier);
+    if (match) {
+        const target = pathResolve(__dirname, '../../../target/generic-jar-content/' + match[1]);
         return {
-            url: pathToFileURL(pathMappings[specifier]).href,
+            url: pathToFileURL(target).href,
             shortCircuit: true
         };
     }
